scipy: isolate from gfortran version
Scipy is dynamically linked to libgfortran, libquadmath, and libgcc.
When gcc is upgraded and the old gcc keg is removed from the cellar,
scipy will break. There is no good version-independent path to these
libs. Traditionally, one might turn to static linking, but
statically linking to libquadmath in particular is difficult in
the context of the numpy distutils build process.

To solve, we borrow matthew-brett/delocate and use it to move the
versions of the gcc libraries we've linked against into the keg's
libexec directory and rewrite each .so to point to the new home,
gaining the benefits of static linkage without having to faff around
with numpy's distutils.